fix: epl -> core 모듈로 이름 변경 및 알림 모듈과 통합
