+++ MSA-core/build.gradle
