+++ MSA-core/src/main/java/com/joh/core/CoreApplication.java
@@ -1,0 +1,17 @@
+package com.joh.core;
+
+import java.util.TimeZone;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication(scanBasePackages = {"com.joh.common", "com.joh.core"})
+@EnableDiscoveryClient
+public class CoreApplication {
+
+  public static void main(String[] args) {
+    TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
+    SpringApplication.run(CoreApplication.class, args);
+  }
+
+}
