+++ MSA-core/src/main/java/com/joh/core/club/model/EplClub.java
@@ -1,0 +1,5 @@
+package com.joh.core.club.model;
+
+public class EplClub {
+
+}
