+++ MSA-core/src/main/java/com/joh/core/club/repository/EplClubRepository.java
@@ -1,0 +1,8 @@
+package com.joh.core.club.repository;
+
+import com.joh.core.club.model.EplClub;
+import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
+
+public interface EplClubRepository extends ReactiveMongoRepository<EplClub, String> {
+
+}
