+++ MSA-core/src/main/java/com/joh/core/reward/model/Reward.java
@@ -1,0 +1,5 @@
+package com.joh.core.reward.model;
+
+public class Reward {
+
+}
