+++ MSA-core/src/main/java/com/joh/core/user/config/WebClientConfig.java
@@ -1,0 +1,14 @@
+package com.joh.core.user.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+public class WebClientConfig {
+
+  @Bean
+  public WebClient webClient() {
+    return WebClient.builder().build();
+  }
+}
