+++ MSA-core/src/main/java/com/joh/core/user/dto/LoginCheckDTO.java
@@ -1,0 +1,19 @@
+package com.joh.core.user.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LoginCheckDTO {
+
+  private boolean isSignUp;
+  private UserResponse userResponse;
+
+  @Builder
+  public LoginCheckDTO(boolean isSignUp, UserResponse userResponse) {
+    this.isSignUp = isSignUp;
+    this.userResponse = userResponse;
+  }
+}
