+++ MSA-core/src/main/java/com/joh/core/user/dto/UserResponse.java
@@ -1,0 +1,21 @@
+package com.joh.core.user.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class UserResponse {
+
+  private String email;
+  private String nickname;
+  private String profileImage;
+
+  @Builder
+  public UserResponse(String email, String nickname, String profileImage) {
+    this.email = email;
+    this.nickname = nickname;
+    this.profileImage = profileImage;
+  }
+}
