+++ MSA-core/src/main/java/com/joh/core/user/exception/UserEmailUpdateException.java
@@ -1,0 +1,8 @@
+package com.joh.core.user.exception;
+
+public class UserEmailUpdateException extends RuntimeException {
+
+  public UserEmailUpdateException(String message) {
+    super(message);
+  }
+}
