+++ MSA-core/src/main/java/com/joh/core/user/exception/UserNotFoundException.java
@@ -1,0 +1,8 @@
+package com.joh.core.user.exception;
+
+public class UserNotFoundException extends RuntimeException {
+
+  public UserNotFoundException(String message) {
+    super(message);
+  }
+}
