+++ MSA-core/src/main/java/com/joh/core/user/exception/UserRegistrationException.java
@@ -1,0 +1,8 @@
+package com.joh.core.user.exception;
+
+public class UserRegistrationException extends RuntimeException {
+
+  public UserRegistrationException(String message) {
+    super(message);
+  }
+}
