+++ MSA-core/src/main/java/com/joh/core/user/exception/kakao/KakaoAuthorizationCodeNullPointerException.java
@@ -1,0 +1,8 @@
+package com.joh.core.user.exception.kakao;
+
+public class KakaoAuthorizationCodeNullPointerException extends RuntimeException {
+
+  public KakaoAuthorizationCodeNullPointerException(String message) {
+    super(message);
+  }
+}
