+++ MSA-core/src/main/java/com/joh/core/user/exception/kakao/KakaoGetUserInfoException.java
@@ -1,0 +1,8 @@
+package com.joh.core.user.exception.kakao;
+
+public class KakaoGetUserInfoException extends RuntimeException {
+
+  public KakaoGetUserInfoException(String message) {
+    super(message);
+  }
+}
