+++ MSA-core/src/main/java/com/joh/core/user/exception/kakao/KakaoUserInfoConvertValueException.java
@@ -1,0 +1,8 @@
+package com.joh.core.user.exception.kakao;
+
+public class KakaoUserInfoConvertValueException extends RuntimeException {
+
+  public KakaoUserInfoConvertValueException(String message) {
+    super(message);
+  }
+}
