+++ MSA-core/src/main/java/com/joh/core/user/model/utils/OAuthProvider.java
@@ -1,0 +1,6 @@
+package com.joh.core.user.model.utils;
+
+public enum OAuthProvider {
+  KAKAO,
+  NAVER;
+}
