+++ MSA-core/src/main/java/com/joh/core/user/repository/CustomUserRepository.java
@@ -1,0 +1,7 @@
+package com.joh.core.user.repository;
+
+import reactor.core.publisher.Mono;
+
+public interface CustomUserRepository {
+  Mono<Long> updateEmailByIdAndNewEmail(String id, String newEmail);
+}
