--- MSA-epl/src/main/java/com/joh/epl/EplApplication.java
@@ -1,17 +1,0 @@
-package com.joh.epl;
-
-import java.util.TimeZone;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication(scanBasePackages = {"com.joh.common", "com.joh.epl"})
-@EnableDiscoveryClient
-public class EplApplication {
-
-  public static void main(String[] args) {
-    TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
-    SpringApplication.run(EplApplication.class, args);
-  }
-
-}
