--- MSA-epl/src/main/java/com/joh/epl/config/WebClientConfig.java
@@ -1,14 +1,0 @@
-package com.joh.epl.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-
-  @Bean
-  public WebClient webClient() {
-    return WebClient.builder().build();
-  }
-}
