--- MSA-epl/src/main/java/com/joh/epl/controller/UserController.java
@@ -1,26 +1,0 @@
-package com.joh.epl.controller;
-
-import java.util.HashMap;
-import java.util.Map;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import reactor.core.publisher.Mono;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/epl")
-public class UserController {
-
-  @GetMapping("/users/test")
-  public Mono<ResponseEntity<Map<String, String>>> test(@RequestHeader("X-USER-ID") String oauthId) {
-
-    Map<String, String> res = new HashMap<>();
-    res.put("oauthId", oauthId);
-
-    return Mono.just(ResponseEntity.ok(res));
-  }
-}
