--- MSA-epl/src/main/java/com/joh/epl/dto/LoginCheckDTO.java
@@ -1,19 +1,0 @@
-package com.joh.epl.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class LoginCheckDTO {
-
-  private boolean isSignUp;
-  private UserResponse userResponse;
-
-  @Builder
-  public LoginCheckDTO(boolean isSignUp, UserResponse userResponse) {
-    this.isSignUp = isSignUp;
-    this.userResponse = userResponse;
-  }
-}
