--- MSA-epl/src/main/java/com/joh/epl/dto/UserResponse.java
@@ -1,21 +1,0 @@
-package com.joh.epl.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class UserResponse {
-
-  private String email;
-  private String nickname;
-  private String profileImage;
-
-  @Builder
-  public UserResponse(String email, String nickname, String profileImage) {
-    this.email = email;
-    this.nickname = nickname;
-    this.profileImage = profileImage;
-  }
-}
