--- MSA-epl/src/main/java/com/joh/epl/exception/UserEmailUpdateException.java
@@ -1,8 +1,0 @@
-package com.joh.epl.exception;
-
-public class UserEmailUpdateException extends RuntimeException {
-
-  public UserEmailUpdateException(String message) {
-    super(message);
-  }
-}
