--- MSA-epl/src/main/java/com/joh/epl/exception/UserNotFoundException.java
@@ -1,8 +1,0 @@
-package com.joh.epl.exception;
-
-public class UserNotFoundException extends RuntimeException {
-
-  public UserNotFoundException(String message) {
-    super(message);
-  }
-}
