--- MSA-epl/src/main/java/com/joh/epl/exception/UserRegistrationException.java
@@ -1,8 +1,0 @@
-package com.joh.epl.exception;
-
-public class UserRegistrationException extends RuntimeException {
-
-  public UserRegistrationException(String message) {
-    super(message);
-  }
-}
