--- MSA-epl/src/main/java/com/joh/epl/exception/kakao/KakaoAuthorizationCodeNullPointerException.java
@@ -1,8 +1,0 @@
-package com.joh.epl.exception.kakao;
-
-public class KakaoAuthorizationCodeNullPointerException extends RuntimeException {
-
-  public KakaoAuthorizationCodeNullPointerException(String message) {
-    super(message);
-  }
-}
