--- MSA-epl/src/main/java/com/joh/epl/exception/kakao/KakaoGetUserInfoException.java
@@ -1,8 +1,0 @@
-package com.joh.epl.exception.kakao;
-
-public class KakaoGetUserInfoException extends RuntimeException {
-
-  public KakaoGetUserInfoException(String message) {
-    super(message);
-  }
-}
