--- MSA-epl/src/main/java/com/joh/epl/exception/kakao/KakaoUserInfoConvertValueException.java
@@ -1,8 +1,0 @@
-package com.joh.epl.exception.kakao;
-
-public class KakaoUserInfoConvertValueException extends RuntimeException {
-
-  public KakaoUserInfoConvertValueException(String message) {
-    super(message);
-  }
-}
