--- MSA-epl/src/main/java/com/joh/epl/model/EplClub.java
@@ -1,5 +1,0 @@
-package com.joh.epl.model;
-
-public class EplClub {
-
-}
