--- MSA-epl/src/main/java/com/joh/epl/model/Reward.java
@@ -1,5 +1,0 @@
-package com.joh.epl.model;
-
-public class Reward {
-
-}
