--- MSA-epl/src/main/java/com/joh/epl/model/utils/OAuthProvider.java
@@ -1,6 +1,0 @@
-package com.joh.epl.model.utils;
-
-public enum OAuthProvider {
-  KAKAO,
-  NAVER;
-}
