--- MSA-epl/src/main/java/com/joh/epl/repository/EplClubRepository.java
@@ -1,8 +1,0 @@
-package com.joh.epl.repository;
-
-import com.joh.epl.entity.EplClub;
-import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
-
-public interface EplClubRepository extends ReactiveMongoRepository<EplClub, String> {
-
-}
