--- MSA-epl/src/main/java/com/joh/epl/repository/RewardRepository.java
@@ -1,8 +1,0 @@
-package com.joh.epl.repository;
-
-import com.joh.epl.entity.Reward;
-import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
-
-public interface RewardRepository extends ReactiveMongoRepository<Reward, String> {
-
-}
