--- MSA-epl/src/main/java/com/joh/epl/repository/user/CustomUserRepository.java
@@ -1,7 +1,0 @@
-package com.joh.epl.repository.user;
-
-import reactor.core.publisher.Mono;
-
-public interface CustomUserRepository {
-  Mono<Long> updateEmailByIdAndNewEmail(String id, String newEmail);
-}
